ComedyCraft - "Random Joke Generator"
ComedyCraft is a Free One-Click Joke Generator that shows you a fresh joke anytime you wish to read one.
The "ComedyCraft" is a fun Chrome extension that delivers a dose of humor with just a click. Get ready to laugh as it presents a collection of random jokes right in your browser, bringing a smile to your face whenever you need it.With 200,000+ funny jokes and a clean UI, JokeBox is the gift you want to give yourself.
Technology used :-
HTML
CSS
JAVASCRIPT